Adding 1-array_iterator.c file

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int it;
+	size_t it;
 
 	if (array == NULL || action == NULL)
 		return;
